Add input buffer Frame rate counter Superjump Attack

+++ src/physics.rs
@@ -1,6 +1,6 @@
 use specs::{Join, System, WriteStorage};
 
-use sm::PhysicsData;
+use crate::PhysicsData;
 
 pub struct Physics;
 
@@ -9,8 +9,8 @@ impl<'a> System<'a> for Physics {
 
     fn run(&mut self, mut data: Self::SystemData) {
         for physics_data in (&mut data).join() {
-            physics_data.h_speed += physics_data.h_acceleration as i32;
-            physics_data.v_speed += physics_data.v_acceleration as i32;
+            physics_data.h_speed += physics_data.h_acceleration;
+            physics_data.v_speed += physics_data.v_acceleration;
             physics_data.position = physics_data.position.offset(physics_data.h_speed, physics_data.v_speed);
         }
     }
